After a lot of hard work, there are results—the website is almost fully implemented for a logged-in user. I’m updating now because I will probably forget everything by morning:
I moved the navigation menu to be handled by JavaScript, and I made it so it recognizes whether a user is logged in or not. If not, it suggests logging in; if yes, it shows the logout button and a profile button. Finally, if the user is an admin, an admin page button will also be displayed.
I added protections on both the client and server sides to prevent access to the admin page unless you are an admin.
I completed the personal page, which currently checks if you are logged in or not. When you log in, you are automatically redirected to it, and the details of the logged-in user are displayed.